uint

+++ models/gas/ethereum/gas_ethereum_fees_traces.sql
@@ -58,7 +58,7 @@ WITH traces AS (
           , CAST(NULL AS varbinary) AS input
           , CAST(NULL AS varchar) AS trace_type
           , CAST(NULL AS varchar) AS call_type
-          , CAST(NULL AS bigint) AS trace_value
+          , CAST(NULL AS uint256) AS trace_value
           , CAST(NULL AS boolean) AS trace_success
           , CAST(NULL AS boolean) AS tx_success
           FROM {{ source('ethereum','traces') }}
